+++ src/main/java/com/example/trial/controller/CountryController.java
@@ -1,10 +1,9 @@
 package com.example.trial.controller;
 
+import com.example.trial.errorhandling.ResourceNotFoundException;
 import com.example.trial.model.Country;
 import com.example.trial.services.DataGenerationService;
-import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -17,12 +16,16 @@ public class CountryController {
 
     @GetMapping
     public List<Country> getAllCountries() {
+        if (dataGenerationService.getAllCountries().isEmpty())
+            throw new ResourceNotFoundException("No countries found");
         return dataGenerationService.getAllCountries();
     }
 
 
     @GetMapping("/{id}")
     public Country getCountryById(@PathVariable String id) {
+        if (dataGenerationService.getCountryById(id)==null)
+            throw new ResourceNotFoundException("Country with iso code "+id+" was not found");
         return dataGenerationService.getCountryById(id);
     }
 
@@ -32,15 +35,18 @@ public class CountryController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Country> updateCountry(@RequestBody Country country,@PathVariable String id) {
-        Country updatedCountry = dataGenerationService.saveCountry(dataGenerationService.getCountryById(id));
-        if (updatedCountry != null) {
-            return ResponseEntity.ok(updatedCountry);
-        }
-        return ResponseEntity.notFound().build();
+    public Country updateCountry(@RequestBody String name,@PathVariable String id) {
+        Country c=dataGenerationService.getCountryById(id);
+        if (c==null)
+            throw new ResourceNotFoundException("No country with iso code "+id);
+        c.setName(name);
+        return dataGenerationService.saveCountry(c);
     }
     @DeleteMapping("/{id}")
     public void deleteCountry(@PathVariable String id) {
+        Country c=dataGenerationService.getCountryById(id);
+        if (c==null)
+            throw new ResourceNotFoundException("No country with iso code "+id);
         dataGenerationService.deleteCountry(id);
     }
 }
